+++ src/main/java/com/ecommerce/database/repository/jpa/OrderItemJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderItemJpaRepository extends JpaRepository<OrderItemEntity, Long> {
+public interface OrderItemJpaRepository extends JpaRepository<OrderItemEntity, Integer> {
 }
