+++ src/main/java/com/ecommerce/database/repository/jpa/ProductJpaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductJpaRepository extends JpaRepository<ProductEntity, Long> {
+public interface ProductJpaRepository extends JpaRepository<ProductEntity, Integer> {
 
     List<ProductEntity> findByName(String name);
 }
